refactor: 도메인 검증 역할 서비스로 위임
